complete refactoring: merge both drag rotate hooks into one
